requirements-dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements-dev.txt requirements-dev.in
#
appdirs==1.4.3            # via setuptools
coverage==4.3.4
django-appconf==1.0.2
flake8==3.3.0
isort==4.2.5
mccabe==0.6.1
packaging==16.8           # via setuptools
pep8-naming==0.4.1
py==1.4.33                # via pytest
pycodestyle==2.3.1        # via flake8
pydocstyle==1.1.1
pyflakes==1.5.0           # via flake8
pyparsing==2.2.0          # via packaging
pytest-django==3.1.2
pytest==3.0.7
selenium==3.3.1
six==1.10.0               # via packaging, setuptools

# The following packages are considered to be unsafe in a requirements file:
# setuptools                # via pytest
